SGB with user-defined loss function

SGB with user-defined loss function.

sgb.loss(
  data.train,
  data.valid,
  data.test,
  otherdata1,
  otherdata2,
  otherdata3,
  otherdata4,
  otherdata5,
  otherdata6,
  y.type,
  pos.y,
  pos.covariates,
  pi,
  pi.valid,
  loss,
  reorder,
  conditional,
  miniter,
  maxiter,
  rho,
  stoch,
  ratio,
  m.convergence.train,
  m.convergence.valid,
  m.convergence.test,
  delta.convergence,
  increment.convergence,
  m0.train,
  m1.train,
  m0.valid,
  m1.valid,
  m0.test,
  m1.test,
  verbose = TRUE
)

Arguments

data.train
training data set

data.valid
validation data set

data.test
test data set

otherdata1
other data used for the reordered classic loss, 1

otherdata2
other data used for the reordered classic loss, 2

otherdata3
other data used for the reordered classic loss, 3

otherdata4
other data used for the reordered classic loss, 4

otherdata5
other data used for the reordered classic loss, 5

otherdata6
other data used for the reordered classic loss, 6

y.type
churn variable, and

pos.y
positions (col) of churn

pos.covariates
covariates

pi
individual expected profit lifts used as DV and weights in the profit loss function

pi.valid
individual expected profit lifts used as DV and weights in the profit loss function

loss
specify loss function and type of offer

reorder
whether scores should be reordered

conditional
whether the offer is conditional on renewal

miniter
miniter hyperparameter

maxiter
maxiter hyperparameter

rho
learning rate

stoch
use stochastic GB or regular GB

ratio
fraction of observations to be sampled in case of stochastic GB

m.convergence.train
training scores to determine convergence by calculating optimal campaign profit at each iteration

m.convergence.valid
validation scores to determine convergence by calculating optimal campaign profit at each iteration

m.convergence.test
test scores to determine convergence by calculating optimal campaign profit at each iteration

delta.convergence
intervention cost to determine convergence by calculating optimal campaign profit at each iteration

increment.convergence
granularity level to determine convergence by calculating optimal campaign profit at each iteration

m0.train
non-targeted margin scores reordered classic loss train sample

m1.train
targeted margin scores reordered classic loss train sample

m0.valid
non-targeted margin scores reordered classic loss validation sample

m1.valid
targeted margin scores reordered classic loss validation sample

m0.test
non-targeted margin scores reordered classic loss test sample

m1.test
targeted margin scores reordered classic loss test sample

verbose
verbose output (default: TRUE)

Value

A list containing best iteration of sgb (best.iteration.valid), scores at best iteration (fmts), variables' importance (Variable.Importance), campaign profit at best target size for each iteration of SGB (maxlift.convergence.valid), target size at best iteration optimized on valid sample (targetsize.convergence.valid), and other SGB outcome measures.

iterations

best.iteration.valid

LogLik

ytrain

yv

yts

fmt

fmv

fmts

Variable.Importance

maxlift.convergence.valid

targetsize.convergence.valid

Details

This function applies the SGB algorithm with a user-defined loss function. It calibrates the model on a calibration sample and generates predictions on a validation and test samples.