Target Size Optimization

Target Size Optimization in RCT.

targetsizeoptimization(
  y,
  treated,
  scores,
  m,
  delta,
  conditional = TRUE,
  increment = 0.1,
  plot = TRUE
)

Arguments

y
post-intervention churn dummy

treated
treatment dummy

scores
predictions

m
post-intervention cash flows

delta
intervention cost

conditional
conditional offer or not (default: TRUE)

increment
share of customers added at a time to build the campaign profit curve (default: .1)

plot
plot the campaign profit curve (default: TRUE)

Value

A list containing campaign profit curve, target sizes, maximum profit and related variables.

campaign.profit.curve

target.sizes

maxprofit

targetsize.maxprofit

maxprofit.trimmed

targetsize.maxprofit.trimmed

secondbesttargetsize.trimmed

increment