W-SGB

Procedure W-SGB Model

wsgb.procedure(
  myseed,
  mydata,
  pos.covariates,
  pos.y,
  delta,
  conditional,
  loss,
  reorder = FALSE,
  pi,
  pi.valid,
  increment,
  maxiter = 3000,
  miniter = 1500,
  rho = 0.001,
  stoch = TRUE,
  ratio = 0.3,
  m0.train,
  m1.train,
  m0.valid,
  m1.valid,
  m0.test,
  m1.test,
  verbose = TRUE
)

Arguments

myseed
seed for reproducibility of the results

mydata
input data

pos.covariates
position (column number of the covariates)

pos.y
indicates the position (column number) of y in the data set

delta
intervention cost

conditional
whether the offer is conditional on renewal

loss
loss function

reorder
whether scores should be reordered (default: FALSE)

pi
individual expected profit lifts used as DV and weights in the profit loss function - not needed for classic loss increment, used to assess convergence in SGB when computing optimal target size

pi.valid
individual expected profit lifts used as DV and weights in the profit loss function - not needed for classic loss increment, used to assess convergence in SGB when computing optimal target size

increment
granularity level to determine optimal campaign size

maxiter
maxiter hyperparameter SGB (default: 3000)

miniter
miniter hyperparameter SGB (default: 1500)

rho
rho hyperparameter SGB (default: .001)

stoch
stoch hyperparameter SGB (default: TRUE)

ratio
ratio hyperparameter SGB (default: .3)

m0.train
nontargeted margin scores reordered classic loss train sample

m1.train
targeted margin scores reordered classic loss train sample

m0.valid
nontargeted margin scores reordered classic loss valid sample

m1.valid
targeted margin scores reordered classic loss valid sample

m0.test
nontargeted margin scores reordered classic loss test sample

m1.test
targeted margin scores reordered classic loss test sample

verbose
verbose output (default: TRUE)

Value

List with W-SGB outcome measures.

myseed

conditional

delta

train.data

valid.data

test.data

scores.train

scores.valid

scores.test

best.iteration.valid

LogLik

Variable.Importance

maxlift.convergence.valid

targetsize.convergence.valid